feat(vision): 新增多任务视觉检测支持并完善全流程
修复多处日志格式化参数不匹配问题，重构在线检测器通信协议并支持任务持久化。
新增连续检测自动调度、跨帧结果缓存，支持LCD叠加显示多任务状态。
新增坐姿异常提醒与多项MCP控制工具，扩展检测结果与性能统计字段。
优化字体绘制与bbox解析逻辑，简化重复代码。

+++ main/vision/face_detector.cc
@@ -116,9 +116,9 @@ DetectionResult FaceDetector::Detect(const ImageFrame& frame) {
     }
 
     auto end = esp_timer_get_time();
-    ESP_LOGD(TAG, "Detected %d faces in %lld ms (threshold=%.2f)",
+    ESP_LOGD(TAG, "Detected %d faces in %d ms (threshold=%.2f)",
              (int)result.detections.size(),
-             (long long)((end - start) / 1000LL),
+             (int)((end - start) / 1000LL),
              confidence_threshold_);
 #else
     (void)frame;
